+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/types/GRADRECORD.java
@@ -18,12 +18,13 @@ package com.jpexs.decompiler.flash.types;
 
 import com.jpexs.decompiler.flash.types.annotations.Internal;
 import com.jpexs.decompiler.flash.types.annotations.SWFType;
+import java.io.Serializable;
 
 /**
  *
  * @author JPEXS
  */
-public class GRADRECORD {
+public class GRADRECORD implements Serializable {
 
     @SWFType(BasicType.UI8)
     public int ratio;